refactor(web): split catalog page routes into focused helpers [OPE-331]
