+++ crates/opengoose-web/src/routes/pages/catalog/pages/mod.rs
@@ -1,0 +1,24 @@
+mod agents;
+mod queue;
+mod runs;
+mod scaffold;
+mod schedules;
+mod sessions;
+mod teams;
+mod triggers;
+mod workflows;
+
+pub(in crate::routes::pages::catalog) use scaffold::render_catalog_spec_page;
+pub(in crate::routes::pages::catalog) use schedules::SchedulesSpec;
+pub(in crate::routes::pages::catalog) use sessions::SessionsSpec;
+pub(in crate::routes::pages::catalog) use teams::TeamsSpec;
+pub(in crate::routes::pages::catalog) use triggers::TriggersSpec;
+
+pub(crate) use agents::agents;
+pub(crate) use queue::queue;
+pub(crate) use runs::runs;
+pub(crate) use schedules::schedules;
+pub(crate) use sessions::{sessions, sessions_events};
+pub(crate) use teams::teams;
+pub(crate) use triggers::triggers;
+pub(crate) use workflows::workflows;
